refactor(ui): make reusable ui components for buttons and modals (TT-1979) (#69)
This commit also refactors the usage of buttons and modals in the app to
use the reusable versions. This makes buttons and modals behave
consistently throughout the application.

It also cleans up the usage of colors, instead opting for using
variables such as primary/secondary/warning/danger etc. from
tailwind/HeroUI for more consistency and easier to identify call to
action.